feat: able to connect with socket through env variable
